# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agentic as a Service (AaaS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: List all runs for the tenant
      security:
        - clerkSession: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, awaiting_approval, completed, failed, cancelled, budget_exceeded]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{runId}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Get run details and steps
      security:
        - clerkSession: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{runId}/approve:
    post:
      operationId: approveRun
      tags: [runs]
      summary: Approve a pending tool call (human-in-the-loop)
      security:
        - clerkSession: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run approved and resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{runId}/cancel:
    post:
      operationId: cancelRun
      tags: [runs]
      summary: Cancel a running agent
      security:
        - clerkSession: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List webhooks
      security:
        - clerkSession: []
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookList"

    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Register a webhook
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /webhooks/{webhookId}:
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Remove a webhook
      security:
        - clerkSession: []
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List scheduled triggers
      security:
        - clerkSession: []
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleList"

    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a scheduled trigger
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedules/{scheduleId}:
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a scheduled trigger
      security:
        - clerkSession: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleRequest"
      responses:
        "200":
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a scheduled trigger
      security:
        - clerkSession: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schedule deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Agent:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        model:
          type: string
        tools:
          type: array
          items: {}
        maxSteps:
          type: integer
        maxBudgetCents:
          type: integer
        approvalMode:
          type: string
          enum: [auto, human_in_loop]
        status:
          type: string
          enum: [active, paused, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - description
        - systemPrompt
        - model
        - tools
        - maxSteps
        - maxBudgetCents
        - approvalMode
        - status
        - createdAt
        - updatedAt
...
    AgentRunStep:
      type: object
      properties:
        stepNumber:
          type: integer
        thought:
          type: ["string", "null"]
        toolCalls:
          type: array
          items: {}
        toolResults:
          type: array
          items: {}
        tokensUsed:
          type: integer
      required:
        - stepNumber
        - tokensUsed
    AgentRun:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        tenantId:
          type: string
        trigger:
          type: string
          enum: [manual, api, schedule, webhook]
        input:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [queued, running, awaiting_approval, completed, failed, cancelled, budget_exceeded]
        output:
          type: ["object", "null"]
          additionalProperties: true
        totalTokens:
          type: integer
        costCents:
          type: integer
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        error:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - tenantId
        - trigger
        - input
        - status
        - totalTokens
        - costCents
        - createdAt
    AgentRunDetail:
      allOf:
        - $ref: "#/components/schemas/AgentRun"
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/AgentRunStep"
            agent:
              $ref: "#/components/schemas/Agent"
          required:
            - steps
            - agent
    AgentRunList:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/AgentRun"
        total:
          type: integer
      required:
        - runs
        - total
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        agentId:
          type: ["string", "null"]
        url:
          type: string
        events:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - url
        - events
        - createdAt
    WebhookList:
      type: object
      properties:
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
      required:
        - webhooks
    CreateWebhookRequest:
      type: object
      properties:
        url:
          type: string
        agentId:
          type: string
        events:
          type: array
          items:
            type: string
        secret:
          type: string
      required:
        - url
        - events
        - secret
    Schedule:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        tenantId:
          type: string
        cronExpression:
          type: string
        inputTemplate:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        nextRunAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - tenantId
        - cronExpression
        - inputTemplate
        - enabled
        - createdAt
        - updatedAt
    ScheduleList:
      type: object
      properties:
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/Schedule"
      required:
        - schedules
    CreateScheduleRequest:
      type: object
      properties:
        agentId:
          type: string
        cronExpression:
          type: string
        inputTemplate:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
      required:
        - agentId
        - cronExpression
    UpdateScheduleRequest:
      type: object
      properties:
        cronExpression:
          type: string
        inputTemplate:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
...