# lib/api-zod/src/generated/types/createWebhookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agentic as a Service (AaaS) API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWebhookRequest {
  url: string;
  agentId?: string;
  events: string[];
  secret: string;
}